cleanup

+++ models/mediaExtra/TMediaDeleteExtra.php
@@ -2,28 +2,69 @@
 
 namespace app\models\mediaExtra;
 
+use app\components\Helpers;
 use Yii;
 use add\models\media;
 
 trait TMediaDeleteExtra {
     /**
      * Marks item for deletion
+     * @param $replace was this file replaced by other
      */
-    public function markDeleted() {
+    public function markDeleted($replace = false) {
         $this->is_deleted = 1;
+        $this->time_deleted = time();
+
+        Helpers::transact(function() use ($replace) {
+            if ($this->save()) {
+                if (method_exists($this, 'afterMediaDelete')) $this->afterMediaDelete($replace);
+            }
+        });
+    }
+
+    public function markDeleted1($replace = false) {
+        $this->is_deleted = 1;
+        $this->time_deleted = time();
+
+        $connection = Yii::$app->db;
+        $transaction = $connection->beginTransaction();
+
+        try {
+            if ($this->save()) {
+                if (method_exists($this, 'afterMediaDelete')) $this->afterMediaDelete($replace);
+            }
+        } catch (\Exception $e) {
+            $transaction->rollback();
+
+            throw $e;
+        }
+
+        $transaction->commit();
+    }
+
+    /**
+     * Recovers item from deleetd state
+     */
+    public function recoverDeleted() {
+        $this->is_deleted = 0;
+        $this->time_deleted = 0;
         $this->save();
     }
+
 
     /**
      * Deletes items marked for deletion
      * @static
      *
      */
-    public static function deleteMarked($maxDelete = 0) {
-        $cr = new CDbCriteria(['condition' =>['is_deleted' => 1]]);
-        if ($maxDelete) $cr->limit = $maxDelete;
+    public static function deleteMarked($maxDelete = 100) {
+        $lifeTime = Helpers::getParam('story/deletedLifetime');
+        if (!$lifeTime) throw new \Exception('No lifetime parameter for media deletion');
 
-        $mediaList = Media::model()->findAllByAttributes($cr);
+        $mediaList = self::find()->where(self::makeCondition([
+                                                                'is_deleted'    => 1,
+                                                                'time_deleted'  => ['<', time() - $lifeTime]]
+                                                            ))->limit($maxDelete)->all();
 
         foreach($mediaList as $item) {
             $item->delete();
@@ -38,27 +79,8 @@ trait TMediaDeleteExtra {
         $this->deleteRecursive($this->_thumbFolder);
         $this->deleteRecursive($this->_storeFolder);
 
-        $connection = Yii::app()->db;
-        $transaction = $connection->beginTransaction();
-
-        try {
-            //TextItem::model()->deleteTargetText($this);
+        Helpers::transact(function() {
             parent::delete();
-        } catch (Exception $e) {
-            $transaction->rollback();
-            throw $e;
-        }
-
-        $transaction->commit();
-    }
-
-    /**
-     * Recursive folder deletion
-     * @param $path
-     * @return bool|null
-     * @throws Exception
-     */
-    function deleteRecursive($path) {
-        $this->deleteRecursive($path);
+        });
     }
 }
